requirements.txt
matplotlib==3.1.2
six==1.13.0
torch==1.3.1
scikit_image==0.16.2
imageio==2.6.1
scipy==1.3.3
numpy==1.15.3
opencv_python==4.1.1.26
pandas==1.0.0
torchvision==0.4.2
Pillow==7.2.1
pydicom==1.4.2
scikit_learn==0.22.2.post1
seaborn==0.10.0
SimpleITK==1.2.4
skimage==0.0
tensorboardX==2.0
toml==0.10.0
xlrd==1.2.0
